moved all the facebook files from public folder to new authentication folder
