+++ app/authentication/facebook_passport.js
