+++ app/authentication/fb_user.js
