+++ public/fb_login.html
@@ -4,7 +4,7 @@
         <title>Facebook Authentication</title>
         <link rel="stylesheet" href="//netdna.bootstrapcdn.com/bootstrap/3.0.2/css/bootstrap.min.css">
         <link rel="stylesheet" href="//netdna.bootstrapcdn.com/font-awesome/4.0.3/css/font-awesome.min.css">
-        <script src='./js/facebook_passport.js'></script>
+        <script src='../app/authentication/facebook_passport.js'></script>
         <style>
             body {
                 padding-top:10vmin;
